+++ src/xyz/oihana/schema/constants/traits/ThesaurusTrait.php
@@ -5,6 +5,7 @@ namespace xyz\oihana\schema\constants\traits;
 use xyz\oihana\schema\constants\traits\thesaurus\CollectionTrait;
 use xyz\oihana\schema\constants\traits\thesaurus\ConceptSchemeTrait;
 use xyz\oihana\schema\constants\traits\thesaurus\ConceptTrait;
+use xyz\oihana\schema\constants\traits\thesaurus\DeliveryMethodTermTrait;
 use xyz\oihana\schema\constants\traits\thesaurus\SkosMappingsTrait;
 use xyz\oihana\schema\constants\traits\thesaurus\SkosNotesTrait;
 use xyz\oihana\schema\constants\traits\thesaurus\ThesaurusSchemeTrait;
@@ -20,12 +21,13 @@ use xyz\oihana\schema\constants\traits\thesaurus\TreeMetricsTrait;
  */
 trait ThesaurusTrait
 {
-    use CollectionTrait      ,
-        ConceptSchemeTrait   ,
-        ConceptTrait         ,
-        SkosMappingsTrait    ,
-        SkosNotesTrait       ,
-        ThesaurusSchemeTrait ,
-        ThesaurusTermTrait   ,
-        TreeMetricsTrait     ;
+    use CollectionTrait         ,
+        ConceptSchemeTrait      ,
+        ConceptTrait            ,
+        DeliveryMethodTermTrait ,
+        SkosMappingsTrait       ,
+        SkosNotesTrait          ,
+        ThesaurusSchemeTrait    ,
+        ThesaurusTermTrait      ,
+        TreeMetricsTrait        ;
 }